requirements.txt
cffi==2.0.0
cryptography==46.0.6
pycparser==3.0
python-dotenv==1.2.2
